[WGRCORE-103] First stage of adding tx type to listtransactions.
There are various options to go about adding this info.
Listtransactions gives a basic overview of transactions that belong to our wallet. It formats each relevant vin and vout for JSON outputting.
We can extend it, or create a similar, additional function.
Alternatively, the Qt has a function that could be refactored to live outsite the qt. Question is: will we import additional Qt nastiness?

For now, I'll add several fields to listtransactions and discuss if it fits the purpose of the HTML5 wallet.